refactor(frontend): split deployments table row
